+++ packages/orq-rc/src/models/components/partdoneevent.ts
@@ -76,7 +76,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kk9gg3caypxcwrw3x4ca14fr"),
+  _id: z.string().default("reasoning_01kkb0p178trm9b6tf140v2hhj"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),