+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -403,7 +403,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KK9GG464A5GRYXP9FR0W5H1K"),
+  _id: z.string().default("01KKB0P285P7QMP52ZSKFHXY2V"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,