+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -118,7 +118,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KK9GG463Y7MTVNR34T4SM44E"),
+  _id: z.string().default("01KKB0P282CCM6VD05D20NFA7M"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,